wfa-google-assinstant/requirements.txt
flask
Flask-RESTful==0.3.5
python-socketio

